change to useStore for saving
