+++ services/ui-src/src/types/states.ts
@@ -1,4 +1,9 @@
-import { ParentPageTemplate, PageData, Report } from "types/report";
+import {
+  ParentPageTemplate,
+  PageData,
+  Report,
+  MeasurePageTemplate,
+} from "types/report";
 import React from "react";
 import { BannerData, ErrorVerbiage, HcbsUser } from "types";
 
@@ -48,4 +53,5 @@ export interface HcbsReportState {
   setModalComponent: (modalComponent: React.ReactFragment) => void;
   setAnswers: (answers: any) => void;
   setMeasure: (cmit: number) => void;
+  setSubstitute: (report: Report, selectMeasure: MeasurePageTemplate) => void;
 }
