+++ services/ui-src/src/utils/state/management/reportState.ts
@@ -1,5 +1,10 @@
 import { HcbsReportState } from "types";
-import { ParentPageTemplate, Report } from "types/report";
+import {
+  isMeasureTemplate,
+  MeasurePageTemplate,
+  ParentPageTemplate,
+  Report,
+} from "types/report";
 import { putReport } from "utils/api/requestMethods/report";
 import { getLocalHourMinuteTime } from "utils";
 
@@ -64,5 +69,29 @@ export const mergeAnswers = (answers: any, state: HcbsReportState) => {
   report.pages[pageIndex] = deepMerge(report.pages[pageIndex], answers);
 
   putReport(report); // Submit to API
+  return { report, lastSavedTime: getLocalHourMinuteTime() };
+};
+
+export const substitute = (
+  report: Report,
+  selectMeasure: MeasurePageTemplate
+) => {
+  const measures = report?.pages.filter((page) =>
+    isMeasureTemplate(page)
+  ) as MeasurePageTemplate[];
+
+  if (selectMeasure) {
+    const substitute = selectMeasure.substitutable?.toString();
+    const measure = measures.find((measure) =>
+      measure.id.includes(substitute!)
+    );
+    if (report && measure) {
+      measure.required = true;
+      selectMeasure.required = false;
+
+      putReport(report);
+    }
+  }
+
   return { report, lastSavedTime: getLocalHourMinuteTime() };
 };
