+++ services/ui-src/src/utils/state/useStore.ts
@@ -8,9 +8,14 @@ import {
   ErrorVerbiage,
   AdminBannerState,
 } from "types";
-import { Report } from "types/report";
+import { MeasurePageTemplate, Report } from "types/report";
 import React from "react";
-import { buildState, mergeAnswers, setPage } from "./management/reportState";
+import {
+  buildState,
+  mergeAnswers,
+  setPage,
+  substitute,
+} from "./management/reportState";
 
 // USER STORE
 const userStore = (set: Function) => ({
@@ -91,6 +96,11 @@ const reportStore = (set: Function): HcbsReportState => ({
     }),
   setMeasure: (cmit: number) => {
     set(() => ({ cmit }), false, { type: "setMeasure" });
+  },
+  setSubstitute: (report: Report, selectMeasure: MeasurePageTemplate) => {
+    set(() => substitute(report, selectMeasure), false, {
+      type: "setSubstitute",
+    });
   },
 });
